proyecto/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.1.8
colorama==0.4.6
fastapi==0.116.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.16.0
httptools==0.7.1
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.47.3
typing_extensions==4.15.0
typing-inspection==0.4.1
uvicorn[standard]==0.35.0
watchfiles==1.1.1
websockets==15.0.1
